requirements.txt
beautifulsoup4==4.8.1
black==19.10b0
dropbox==9.4.0
flake8==3.7.9
isort==4.3.21
lxml==4.2.5
mega.py==1.0.5
mypy==0.750
mypy-extensions==0.4.3
pcloud==1.0a6
Pillow==6.2.1
PyQt5==5.13.2
pytest==5.3.1
pytest-cov==2.8.1
reportlab==3.5.23
requests==2.22.0
tabulate==0.8.1
